requirements.txt
torch == 1.10.2
torchvision == 0.11.3
nltk==3.6.7
mysql-connector-python==8.0.28

Flask==1.1.1
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Werkzeug==0.15.5